updated component documenation
