+++ bears-app/src/components/shared/EmailButton/__tests__/index.spec.js
@@ -1,9 +1,9 @@
 import renderer from 'react-test-renderer'
-import Email from '../index.jsx'
+import EmailButton from '../index.jsx'
 
 describe('Email', () => {
   it('renders a match to the previous snapshot', () => {
-    const component = renderer.create(<Email />)
+    const component = renderer.create(<EmailButton />)
     expect(component.toJSON()).toMatchSnapshot()
   })
 })
