+++ bears-app/src/components/shared/index.js
@@ -1,5 +1,6 @@
 // order exports in alphabetical list
 export { default as Button } from './Button'
+export { default as EmailButton } from './EmailButton'
 export { default as Form } from './Form'
 export { default as Heading } from './Heading'
 export { default as Paragraph } from './Paragraph'
